Task_Project/requirements.txt
Django==4.2.1
psycopg2==2.9.6
dj-database-url==2.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djangorestframework-simplejwt[blacklist]
